requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
seaborn
opencv-python
